ielts_requirements.txt
streamlit
pymupdf
pypdf2
python-docx
nltk
language-tool-python
regex